+++ app/src/main/java/cz/darken/app/MainActivityHolder.kt
@@ -21,4 +21,10 @@ object MainActivityHolder {
             activityRef?.get()?.exitAppFully()
         }
     }
+
+    fun notifyOverlayStopped() {
+        activityRef?.get()?.runOnUiThread {
+            activityRef?.get()?.onOverlayStoppedExternally()
+        }
+    }
 }